[AntarexClavaApi] Adds workaround to import bug in code gen library
The library MargotCodeGen can now be imported from LARA aspects weaved by Clava. This commit splits the JS and LARA parts of the library in different fiels, since the JS functions did not recognize the LARA aspects they were supposed to call. The LARA aspects were moved to a separate file, which is imported in the JS file.