[Multiversioning] Updates strategy with code generation options
It is now possible to control if the clones are generated next to the original function or in new files. It is also possible to choose whether the optimization options are applied inside the code with pragmas or at build time. If the latter, a mapping text file is generated. If this option is chosen, clones are automatically generated in separate files.