[LibVersioningCompiler] Fixes several bugs with LibVC
Fixes bug with code generation of the options list.

Generates missing <iostream> header in the code and a private JS function to help with this.

Fixes bug 'calculating' the file path to the file to be compiled. This resulted in the woven version compiling the original kernel file, which did not have C linkage. Now it correclty compiles the woven kernel file and the symbol is loaded as intended.