[mARGOt] Changes config, dse and codegen interaction
Previously, one could create instances of either codegen or dse from config. This meant that config needed to be aware of both of those classes. This would result into importing dse classes when one wanted to just use the config. The current dse implementation relies on LAT and this would have to be added to the weaver configuration just to use the config features.

With this commit, the code that uses the information from config instances to generate either codegen or dse instances, is located inside those respective classes. This means that config knows nothing about the existance of those classes and doesn't import other code that cause dependency problems.